\name{N-Monthly Datings}
\alias{Quarterly}
\alias{HalfYearly}
\docType{data}
\title{N-Monthly Datings}
\description{
The dates of N-Monthly datings are a subset of the Monthly dates with a fixed periodicity (N) from January onwards.

Defined N-Monthly Datings:
\itemize{
\item \code{Quarterly}: The dating containing all the first day of a quarter or trimester. I. e. 1st january, 1st April, 1st July and 1st October.
\item \code{HalfYearly}: The dating containing all the first day of a half-year.  I. e. 1st January and 1st July.
}
}
\seealso{
See also the standard datings \code{\link{Monthly}} and \code{\link{Yearly}}
}