\name{Standard Datings}
\alias{Daily}
\alias{Weekly}
\alias{Monthly}
\alias{Yearly}
\docType{data}
\title{
Standard Datings
}
\description{
Description of the implemented standard datings.

Standard Datings:
\itemize{
\item \code{Daily}: The dating containing all the dates.
\item \code{Weekly}: The dating containing all the sundays (first day of a week in \acronym{R}).
\item \code{Monthly}: The dating containing all the first day of a month.
\item \code{Yearly}: The dating containing all the first day of a year.
}
}