\name{Weekdays Datings}
\alias{Mondays}
\alias{Tuesdays}
\alias{Wednesdays}
\alias{Thursdays}
\alias{Fridays}
\alias{Saturdays}
\alias{Sundays}
\docType{data}
\title{
Weekdays Datings
}
\description{
Description of the weekdays datings.

Weekdays Datings:
\itemize{
\item \code{Mondays}: The dating containing all the mondays.
\item \code{Tuesdays}: The dating containing all the tuesdays.
\item \code{Wednesdays}: The dating containing all the wednesdays.
\item \code{Thursdays}: The dating containing all the thursdays.
\item \code{Fridays}: The dating containing all the fridays.
\item \code{Saturdays}: The dating containing all the saturdays.
\item \code{Sundays}: The dating containing all the sundays. Is the same dating as \code{Weekly}.
}
}
\seealso{
See also the standard dating \code{\link{Weekly}}
}