\name{Lag Operator}
\alias{A}
\alias{B}
\docType{data}
\title{Lag Operators}
\description{
Lag operator or backward operator: \code{B}.

Inverse lag operator or forward operator: \code{A}.
}
\seealso{
See the reference to \link{Polyn} class.
}
\examples{
p <- 1 + B
}
\keyword{lag operator}