\name{tolBasis-package}
\alias{tolBasis-package}
\alias{tolBasis}
\docType{package}
\title{
TOL Basis
}
\description{
\code{tolBasis} provides basis for the integration of TOL in R.
}
\details{
\tabular{ll}{
Package: \tab tolBasis\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-03-02\cr
License: \tab GPL-3\cr
}
}
\seealso{
See the main classes implemented:
\itemize{
\item{\link{Date}: the R class Date used for TOL integration}
\item{\link{Dating}: a practical version of the TOL grammar TimeSet}
\item{\link{Polyn}: an implementation of the TOL grammar Polyn}
\item{\link{Ratio}: an implementation of the TOL grammar Ratio}
\item{\link{Serie}: a limited implementation of the TOL grammar Serie}
}
}
\author{
Pedro Gea

Maintainer: Pedro Gea <pgea@bayesforecast.com>
}
\references{
See the main page of the TOL-Project: \url{https://www.tol-project.org}
}
\keyword{package}
\keyword{tol}
